Initial commit of toolchain, test, and 'skeleton'.
The skeleton is very lacking. TODO: program an LP to
figure out some parameters for kernel configuration
for work stealing.

Something to configure still is google bench to allow
for measurements of performance data.

The toolchain only allows for compilation with clang at
the moment. Need to investigate the performance
difference with NVCC.